Read all sheets from all excel files and save into CSV files

Following the same principle of read_all_sheets_then_save_csv read_all_xlsx_then_save_csv() just loop read_all_sheets_then_save_csv() over all available xlsx files

read_all_xlsx_then_save_csv(path_to_xlsx)

Arguments

path_to_xlsx

a character vector with the path to excel file

Value

A list (one element per .xlsx file found under path_to_xlsx), each the list returned by read_all_sheets_then_save_csv() for that workbook.

Acknowledgment

See: read_sheet_then_save_csv

Examples

# \donttest{
# read and into a csv
xlsx_dir <- system.file("xlsx-examples", package = "misc")
read_all_xlsx_then_save_csv(
  path_to_xlsx = xlsx_dir
)
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_001/mtcars_workbook_001_mtcars_sheet_010.csv saved!
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_002/mtcars_workbook_002_mtcars_sheet_010.csv saved!
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_003/mtcars_workbook_003_mtcars_sheet_010.csv saved!
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_004/mtcars_workbook_004_mtcars_sheet_010.csv saved!
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_005/mtcars_workbook_005_mtcars_sheet_010.csv saved!
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_006/mtcars_workbook_006_mtcars_sheet_010.csv saved!
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_007/mtcars_workbook_007_mtcars_sheet_010.csv saved!
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_008/mtcars_workbook_008_mtcars_sheet_010.csv saved!
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_009/mtcars_workbook_009_mtcars_sheet_010.csv saved!
#> • Creating /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 created!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_001.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_001.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_002.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_002.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_003.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_003.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_004.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_004.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_005.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_005.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_006.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_006.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_007.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_007.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_008.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_008.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_009.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_009.csv saved!
#> ℹ Directory /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010 already exists!
#> • Saving /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_010.csv...
#> ✔ /home/runner/work/misc/misc/data/temp/extracted_sheets/mtcars_workbook_010/mtcars_workbook_010_mtcars_sheet_010.csv saved!
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_001.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_002.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_003.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_004.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_005.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_006.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_007.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_008.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_009.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_001
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_002
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_003
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_004
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_005
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_006
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_007
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_008
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_009
#> NULL
#> 
#> $`/home/runner/work/_temp/Library/misc/xlsx-examples/mtcars_workbook_010.xlsx`$mtcars_sheet_010
#> NULL
#> 
#> 
# }