Add a gitignore file to the project root

add_gitignore() fetch files using the API from gitignore.io. Also, add_gitignore() include tags (created by ctags) into the gitignore file.

add_gitignore(type = "r")

Arguments

type
a character vector with the language to be ignored

Acknowledgment

add_gitignore() is inspired by gitignore::gi_fetch_templates

Examples

if (FALSE) {
# if your project does not have a gitignore, get it with:
add_gitignore()
# want to ignore python files too:
add_gitignore(type = c("r", "python"))
}